fix(verification): apply round-1 review fixes (logging, timeouts, splits)
- Remove @log_function_call from _get_config / _run / _run_with_input /
  verify_git. The decorator logs argument & return values at debug level
  (json.dumps in mcp_coder_utils.log_utils._log_call_start and
  _log_call_success) and would leak user.signingkey, signed payload bytes,
  and the full result dict. Replace with targeted logger.debug calls that
  log only structural metadata (key names, binary paths, overall_ok),
  never values. Honours Decision #11 (no signing-key/payload logging).
- Wrap subprocess.TimeoutExpired around the three direct subprocess paths
  that previously propagated it: signing_binary --version,
  agent_reachable gpg-connect-agent /bye, actual_signature gpg
  --clearsign. Each emits a structured CheckResult instead of raising.
- Add logger.debug on the verify_head outer exception (Decision #8).
- Split verification.py: extract pure helpers to _signing_helpers.py
  (no subprocess, no .importlinter change). verification.py: 604 -> 565.
- Split test_verification.py: move Tier 1 signing-detection tests to
  test_verification_tier1_signing.py. test_verification.py: 620 -> 305.
- Add tests for each new timeout branch and for the no-key-leak
  invariant on _get_config.
